+++ miniapp/src/main/java/com/rakuten/tech/mobile/miniapp/permission/MiniAppCustomPermissionWindow.kt
@@ -6,45 +6,34 @@ import android.view.View
 import androidx.appcompat.app.AlertDialog
 import androidx.recyclerview.widget.DividerItemDecoration
 import androidx.recyclerview.widget.LinearLayoutManager
-import com.rakuten.tech.mobile.miniapp.MiniApp
 import com.rakuten.tech.mobile.miniapp.R
 import kotlinx.android.synthetic.main.window_custom_permission.view.*
 
-// TODO
-class MiniAppCustomPermissionWindow(
-    private val miniApp: MiniApp,
-    private val miniAppId: String,
-    private val activity: Activity
+/**
+ * A class to show default custom permissions UI to manage permissions in this SDK.
+ */
+internal class MiniAppCustomPermissionWindow(
+    private val customPermissionCache: MiniAppCustomPermissionCache
 ) {
-
     private lateinit var customPermissionAlertDialog: AlertDialog
     private lateinit var customPermissionAdapter: MiniAppCustomPermissionAdapter
     private lateinit var customPermissionLayout: View
 
-    private fun getCachedList() = miniApp.getCustomPermissions(miniAppId).pairValues
-
-    private fun getDeniedPermissions(permissionsWithDescription: List<Pair<MiniAppCustomPermissionType, String>>) =
-        permissionsWithDescription.filter { (first) ->
-
-            getCachedList().find {
-                it.first == first && it.second == MiniAppCustomPermissionResult.DENIED
-            } != null
-        }
-
     fun show(
+        activity: Activity,
+        miniAppId: String,
         permissionsWithDescription: List<Pair<MiniAppCustomPermissionType, String>>,
         callback: (List<Pair<MiniAppCustomPermissionType, MiniAppCustomPermissionResult>>) -> Unit
     ) {
         if (miniAppId.isEmpty())
             return
 
-        val deniedPermissions = getDeniedPermissions(permissionsWithDescription)
+        val deniedPermissions = getDeniedPermissions(miniAppId, permissionsWithDescription)
 
         // show permission default UI if there is any denied permission
         if (deniedPermissions.isNotEmpty()) {
-
             // initialize permission view after ensuring if there is any denied permission
-            initPermissionView(activity)
+            initDefaultWindow(activity)
 
             // prepare data for adapter
             val namesForAdapter: ArrayList<MiniAppCustomPermissionType> = arrayListOf()
@@ -60,49 +49,57 @@ class MiniAppCustomPermissionWindow(
                 resultsForAdapter,
                 descriptionForAdapter
             )
-            customPermissionLayout.listCustomPermission.adapter = customPermissionAdapter
 
             // add action listeners
-            addPermissionClickListeners(callback)
+            addPermissionClickListeners(miniAppId, callback)
 
             // preview dialog
             customPermissionAlertDialog.show()
         } else {
-            callback.invoke(getCachedList())
+            callback.invoke(getCachedList(miniAppId))
         }
     }
 
-    private fun initPermissionView(activity: Activity) {
-        customPermissionLayout = getPermissionLayout(activity)
+    private fun initDefaultWindow(activity: Activity) {
+        val layoutInflater = LayoutInflater.from(activity)
+        customPermissionLayout = layoutInflater.inflate(R.layout.window_custom_permission, null)
+        customPermissionLayout.listCustomPermission.layoutManager = LinearLayoutManager(activity)
+        customPermissionLayout.listCustomPermission.addItemDecoration(
+            DividerItemDecoration(activity, DividerItemDecoration.VERTICAL)
+        )
+
         customPermissionAdapter = MiniAppCustomPermissionAdapter()
+        customPermissionLayout.listCustomPermission.adapter = customPermissionAdapter
+
         customPermissionAlertDialog =
-            AlertDialog.Builder(activity, R.style.AppTheme_CustomPermissionDialog).apply {
-                setView(customPermissionLayout)
-            }.create()
+            AlertDialog.Builder(activity, R.style.AppTheme_CustomPermissionDialog).create()
+        customPermissionAlertDialog.setView(customPermissionLayout)
     }
 
     private fun addPermissionClickListeners(
+        miniAppId: String,
         callback: (List<Pair<MiniAppCustomPermissionType, MiniAppCustomPermissionResult>>) -> Unit
     ) {
-        customPermissionLayout.permissionAllow.setOnClickListener {
+        customPermissionLayout.permissionSave.setOnClickListener {
             callback.invoke(customPermissionAdapter.permissionPairs)
             customPermissionAlertDialog.dismiss()
         }
 
-        customPermissionLayout.permissionDontAllow.setOnClickListener {
-            callback.invoke(getCachedList())
+        customPermissionLayout.permissionCloseWindow.setOnClickListener {
+            callback.invoke(getCachedList(miniAppId))
             customPermissionAlertDialog.dismiss()
         }
     }
 
-    private fun getPermissionLayout(activity: Activity): View {
-        val layoutInflater = LayoutInflater.from(activity)
-        val permissionLayout = layoutInflater.inflate(R.layout.window_custom_permission, null)
-        permissionLayout.listCustomPermission.layoutManager = LinearLayoutManager(activity)
-        permissionLayout.listCustomPermission.addItemDecoration(
-            DividerItemDecoration(activity, DividerItemDecoration.VERTICAL)
-        )
+    private fun getCachedList(miniAppId: String) =
+        customPermissionCache.readPermissions(miniAppId).pairValues
 
-        return permissionLayout
+    private fun getDeniedPermissions(
+        miniAppId: String,
+        permissionsWithDescription: List<Pair<MiniAppCustomPermissionType, String>>
+    ) = permissionsWithDescription.filter { (first) ->
+        getCachedList(miniAppId).find {
+            it.first == first && it.second == MiniAppCustomPermissionResult.DENIED
+        } != null
     }
 }